Moved over to local copy of the new std.process.
This was a bit tricky because all the linux-specific stuff like
chaning commands, redirection and setting environent variables on the
command-line don't work with the new process module.

This change is a step towards making bob work under windows and OSX.

The big remaining change re portability is to make the signal handling
code portable.

Also, there is a bug in the signal handling area re shutting down the
reently introduced thread that handles the signals. The whole area needs
rewriting.