Various minor tidyups, plus removal of some blatant Windows porting issues.
Windows doesn't seem to have any equivalent of kill(pid, sig), so
Windows users will have to wait for spawned processes to complete
when there is a build failure or you Ctrl-C bob.

This shouldn't usually be a problem because most compilers are well behaved.

The remaining Windows porting issues are mostly in bob_config.d.